requirements.txt
deta==1.1.0
matplotlib==3.5.2
numpy==1.22.3
pandas==1.4.2
plotly==5.7.0
python-dotenv==1.0.0
scikit_learn==1.1.1
seaborn==0.11.2
streamlit==1.9.0
streamlit_authenticator==0.1.5
